conversion warning in AI fixed

+++ Extinguish/Extinguish/AI.cpp
@@ -136,7 +136,7 @@ void AI::ReInit(vector<unsigned int> team1, vector<unsigned int> team2)
 
 	if (me->GetTag() == "Team1")
 	{
-		fakeTeam = team1.size() - 1;
+		fakeTeam = (int)team1.size() - 1;
 
 		// for each 'person' on team1
 		for (int j = 0; j < team1.size(); ++j)
@@ -154,7 +154,7 @@ void AI::ReInit(vector<unsigned int> team1, vector<unsigned int> team2)
 
 	else if (me->GetTag() == "Team2")
 	{
-		fakeTeam = team2.size() - 1;
+		fakeTeam = (int)team2.size() - 1;
 
 		for (int j = 0; j < team2.size(); ++j)
 		{
